modification de persistence pour utiliser MYSQL ( sauvegrade de l'ancien dans entities(dossier hors des sources) )

+++ jboss-javaee6-webapp/src/main/java/com/gestionnaire/entities/Auteur.java
@@ -18,7 +18,7 @@ public class Auteur implements Serializable {
 	
 	
 	@Id
-	private int idAuteur;
+	private Long idAuteur;
 	private String nom;
 	private String prenom;
