+++ jboss-javaee6-webapp/src/main/java/com/gestionnaire/entities/Categories.java
@@ -17,7 +17,7 @@ public class Categories implements Serializable {
 
 
 	@Id
-	private int idCategories;
+	private Long idCategories;
 	private String nom;
 	
 	@ManyToMany(mappedBy="categories")
