+++ jboss-javaee6-webapp/src/main/java/com/gestionnaire/entities/Enchere.java
@@ -23,7 +23,7 @@ public class Enchere implements Serializable {
 	
 
 	@Id
-	private int idEnchere;
+	private Long idEnchere;
 	
 	@ManyToOne(fetch=FetchType.LAZY)
 	@JoinColumn(name="idLivre")
